refactor(cloud): isolate migration SQL dialect
